+++ wowup-electron/src/app/components/wow-client-options/wow-client-options.component.ts
@@ -22,11 +22,11 @@ import { WarcraftService } from "../../services/warcraft/warcraft.service";
   styleUrls: ["./wow-client-options.component.scss"],
 })
 export class WowClientOptionsComponent implements OnInit, OnDestroy {
-  @Input("installationId") public installationId: string;
+  @Input("installationId") public installationId = "";
 
   private readonly _editModeSrc = new BehaviorSubject(false);
   private readonly _isBusySrc = new BehaviorSubject(false);
-  private installation: WowInstallation;
+  private installation: WowInstallation | undefined;
   private subscriptions: Subscription[] = [];
 
   public readonly addonChannelInfos: {
@@ -34,15 +34,15 @@ export class WowClientOptionsComponent implements OnInit, OnDestroy {
     name: string;
   }[];
 
-  public clientTypeName: string;
-  public clientFolderName: string;
-  public clientLocation: string;
-  public installationModel: WowInstallation;
-  public selectedAddonChannelType: AddonChannelType;
+  public clientTypeName = "";
+  public clientFolderName = "";
+  public clientLocation = "";
+  public installationModel!: WowInstallation;
+  public selectedAddonChannelType: AddonChannelType = AddonChannelType.Stable;
   public editMode$ = this._editModeSrc.asObservable();
   public isBusy$ = this._isBusySrc.asObservable();
 
-  public clientAutoUpdate: boolean;
+  public clientAutoUpdate = false;
 
   public set isBusy(enabled: boolean) {
     this._isBusySrc.next(enabled);
@@ -63,10 +63,18 @@ export class WowClientOptionsComponent implements OnInit, OnDestroy {
   }
 
   public get executableName(): string {
+    if (!this.installation) {
+      return "";
+    }
+
     return this._warcraftService.getExecutableName(this.installation.clientType);
   }
 
   public get wowLogoImage(): string {
+    if (!this.installation) {
+      return "";
+    }
+
     switch (this.installation.clientType) {
       case WowClientType.Beta:
       case WowClientType.ClassicBeta:
@@ -114,25 +122,40 @@ export class WowClientOptionsComponent implements OnInit, OnDestroy {
   }
 
   public onDefaultAddonChannelChange(evt: MatSelectChange): void {
+    if (!this.installationModel) {
+      return;
+    }
     this.installationModel.defaultAddonChannelType = evt.value;
   }
 
   public onDefaultAutoUpdateChange(evt: MatSlideToggleChange): void {
+    if (!this.installationModel) {
+      return;
+    }
+
     this.installationModel.defaultAutoUpdate = evt.checked;
   }
 
   public onClickCancel(): void {
-    this.installationModel = { ...this.installation };
+    if (this.installation) {
+      this.installationModel = { ...this.installation };
+    }
     this.editMode = false;
   }
 
   public onClickSave(): void {
+    if (!this.installationModel) {
+      return;
+    }
+
     this.isBusy = true;
     try {
       // const saveAutoUpdate = this.installationModel.defaultAutoUpdate !== this.installation.defaultAutoUpdate;
 
       this.installation = { ...this.installationModel };
-      this._warcraftInstallationService.updateWowInstallation(this.installation);
+      if (this.installation) {
+        this._warcraftInstallationService.updateWowInstallation(this.installation);
+      }
 
       // if (saveAutoUpdate) {
       //   await this._addonService.setInstallationAutoUpdate(this.installation);
@@ -151,7 +174,7 @@ export class WowClientOptionsComponent implements OnInit, OnDestroy {
       data: {
         title: this._translateService.instant("PAGES.OPTIONS.WOW.CLEAR_INSTALL_LOCATION_DIALOG.TITLE"),
         message: this._translateService.instant("PAGES.OPTIONS.WOW.CLEAR_INSTALL_LOCATION_DIALOG.MESSAGE", {
-          location: this._translateService.instant(this.installation.location),
+          location: this._translateService.instant(this.installation?.location ?? ""),
         }),
       },
     });
@@ -164,14 +187,16 @@ export class WowClientOptionsComponent implements OnInit, OnDestroy {
             return;
           }
 
-          this._warcraftInstallationService.removeWowInstallation(this.installation);
+          if (this.installation) {
+            this._warcraftInstallationService.removeWowInstallation(this.installation);
+          }
         })
       )
       .subscribe();
   }
 
-  private getAddonChannelInfos() {
-    return getEnumList(AddonChannelType).map((type: AddonChannelType) => {
+  private getAddonChannelInfos(): { type: AddonChannelType; name: string }[] {
+    return getEnumList(AddonChannelType).map((type: any) => {
       const channelName = getEnumName(AddonChannelType, type).toUpperCase();
       return {
         type: type,
